+++ templates-partials/members-list.php
@@ -663,8 +663,7 @@ $no_members_message = __('No members found.', 'wicket-acc');
                     </span>
                     <span data-class_wt_hidden="$currentRemoveMemberName !== ''">
                         <?php echo esc_html__('Are you sure you want to remove', 'wicket-acc'); ?>
-                        <span data-text="$currentRemoveMemberName"></span>
-                        <?php echo esc_html__('from this organization?', 'wicket-acc'); ?>
+                        <span data-text="$currentRemoveMemberName"></span>&nbsp;<?php echo esc_html__('from this organization?', 'wicket-acc'); ?>
                     </span>
                     <br>
                     <?php esc_html_e('This action cannot be undone.', 'wicket-acc'); ?>